Move animators to new package
