+++ previewseekbar-exoplayer/src/main/java/com/github/rubensousa/previewseekbar/exoplayer/PreviewTimeBar.java
@@ -28,11 +28,11 @@ import android.widget.FrameLayout;
 import androidx.annotation.NonNull;
 import androidx.core.content.ContextCompat;
 
-import com.github.rubensousa.previewseekbar.PreviewAnimator;
 import com.github.rubensousa.previewseekbar.PreviewBar;
 import com.github.rubensousa.previewseekbar.PreviewDelegate;
 import com.github.rubensousa.previewseekbar.PreviewLoader;
 import com.github.rubensousa.previewseekbar.PreviewSeekBar;
+import com.github.rubensousa.previewseekbar.animator.PreviewAnimator;
 import com.google.android.exoplayer2.ui.DefaultTimeBar;
 import com.google.android.exoplayer2.ui.TimeBar;
